pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"


[project]
name = "quickstatements"
version = "3.0.0"
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django==5.0.9",
    "uwsgi==2.0.26",
    "requests==2.32.3",
    "requests-mock==1.10.0",
    "authlib==1.3.1",
    "python-dotenv==1.0.1",
    "djangorestframework==3.15.2",
    "markdown==3.7",
    "django-filter==24.3",
    "jsonpatch==1.33",
    "uvicorn (>=0.34.0,<0.35.0)"
]

[tool.pycodestyle]
ignore = "E731,E203,E402,W503,W504,C814"
max-line-length = 99

[tool.isort]
line_length = 99
known_first_party = "quickstatements"
profile = "black"

[tool.black]
line-length = 99
target-version = ['py310', 'py311', 'py312', 'py313']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | build
  | dist
)/
'''
